requirements.txt
altair==5.4.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2026.4.22
cffi==1.17.1
charset-normalizer==3.4.7
click==8.1.8
cryptography==47.0.0
gitdb==4.0.12
GitPython==3.1.50
google-ai-generativelanguage==0.1.0
google-api-core==2.29.0
google-auth==2.50.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.73.0
grpcio==1.70.0
grpcio-status==1.62.3
idna==3.13
importlib_resources==6.4.5
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.42.1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
proto-plus==1.27.1
protobuf==4.25.9
pyarrow==17.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.2
referencing==0.35.1
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.3
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2026.2
urllib3==2.2.3
zipp==3.20.2
